Purchase + PurchaseNotice
Aggiunti Dao, Service e Controller
Da fare i test

+++ src/main/java/it/unical/asd/group6/computerSparePartsCompany/data/controller/ProductController.java
@@ -33,7 +33,7 @@ public class ProductController {
     }
 
     @GetMapping("/showallbypriceislessthan")
-    public ResponseEntity<List<Product>> showAllByByPriceIsLessThan(@RequestParam("price") Integer price){
+    public ResponseEntity<List<Product>> showAllByByPriceIsLessThan(@RequestParam("price") Double price){
         return ResponseEntity.ok(productService.getAllProductByPriceIsLessThan(price));
     }
 }
